% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antminer.R
\name{antminer}
\alias{antminer}
\title{Antminer}
\usage{
antminer(data, class, maxUncoveredCases, NumberOfAnts, NumberOfRulesConverge,
  MinCasesPerRule)
}
\arguments{
\item{data}{zbior danych do budowy modelu}

\item{class}{nazwa kolumny ze zbioru danych z wartosciami klas.
Algorytm zbuduje model do predykcji tych klas}

\item{maxUncoveredCases}{liczba niepokrytych przypadkow ze zbioru danych.
Algorytm bedzie pracowal dopoki liczba niepokrytych przypadkow jest wieksza od tej liczby}

\item{NumberOfAnts}{liczba tworzonych regul podczas jednej iteracji.}

\item{NumberOfRulesConverge}{liczba stworzonych regul podczas iteracji, ktore sie powtarzaja.
Algorytm konczy iteracje powyzej tej liczby.}

\item{MinCasesPerRule}{liczba pokrytych przypadklw przez zbudowana regule.
Jesli regula pokrywa mniej to jest odrzucana.}
}
\value{
Model skladajacy sie z listy abudowanych regul
}
\description{
Algorytm do budowy modelu klasyfikacji oparty na algorytmie mrowkowym optymalizacji
}
\details{
Algorytm w kazdej iteracji tworzy maksymalnie NumberOfAnts regul na podstawie aktualnego zbioru danych.
Regula jest dodawana do zbioru regul jesli pokrywa co najmniej MinCasesPerRule przypadkow.
Ze zbioru stworzonych regul wybierana jest jedna regula, najlepsza pod wzgledem jakosci. Przypadki pokrywane przez ta regule sa usuwane ze zbioru danych.
Algorytm moze wczesniej zakonczyc iteracje jesli liczba zduplikowanych regul ktore stworzyl przekroczy wartosc NumberOfRulesConverge.
Jesli algorytm ciagle tworzy zduplikowane reguly to znaczy ze tak juz sie dostosowal i nie ma sensu tworzyc kolejnych regul, zatem konczymy iteracje.
Algorytm bedzie wykonywal kolejne iteracje jesli liczba przypadkow ze zbioru danych jest wieksza niz maxUncoveredCases.
Stworzony model ma forme listy uporzadkowanych regul. Od pierwszej znalezionej reguly do ostatniej znalezionej reguly.
Wnioskowanie klasy na podstawie modelu polega na znalezieniu regyly ktora pokrywa przypadek i ustawieniu klasy z tego przypadku.
Wazne jest to, ze pierwsza regula ktora pokrywa przyklad jest wybierana, nie sa przegladane inne reguly ktore moga lepiej pokrywac przyklad.
W modelu ustawiana jest takze klasa domyslna, jest to dominujaca klasa z niepokrytych przypadkow.
}
\examples{
model <- antminer(trainingSet, "Class", 10, 100, 15, 1)
pred <- predict(model, testSet)
}

